1.18wmf1 REL r100402, r100405, r100419, r100427, r100432, r100433, r100435, r100436, r100437, r100428, r100439
git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/branches/wmf@100440 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ 1.18wmf1/extensions/Contest/specials/SpecialEditContest.php
@@ -95,6 +95,13 @@ class SpecialEditContest extends FormSpecialPage {
 	protected function showContent( $subPage ) {
 		$isNew = $this->getRequest()->wasPosted() && $this->getUser()->matchEditToken( $this->getRequest()->getVal( 'newEditToken' ) );
 
+		$this->getOutput()->addScript(
+			Skin::makeVariablesScript(
+				array(
+					'ContestDeletionEnabled' => ContestSettings::get( 'contestDeletionEnabled' ),
+				)
+			)
+		);
 		if ( $isNew ) {
 			$data = array( 'name' => $this->getRequest()->getVal( 'newcontest' ) );
 
@@ -312,7 +319,12 @@ class SpecialEditContest extends FormSpecialPage {
 		if ( $idString == '' ) {
 			return true;
 		}
-
+		
+		if ( !ContestSettings::get( 'contestDeletionEnabled' ) ) {
+			// Shouldn't get here (UI should prevent it)
+			throw new MWException( 'Contest deletion is disabled', 'contestdeletiondisabled' );
+		}
+		
 		return ContestChallenge::s()->delete( array( 'id' => explode( '|', $idString ) ) );
 	}
